use headlines with :detail: tag for <detail> export
The headline text is exported as the <summary> in the <detail>.

Using headlines fixes the problem with nested paragraphs, they are now
exported correctly.

On the downside, non-empty headline texts don't work, so a <summary>
must always be exported.

Also the only way to end a headline section is by giving the next
headline, so a <detail> can only come right before and up to the next
headline.