Summary: work around too strict package-lint check
## background information

package-lint tries to flag anything that can be a problem with MELPA
package.  It is not intended to be a generic CI tool and thus it can't
(yet) skip single issues.
see https://github.com/purcell/package-lint/issues/207

## what happened

The monthly ox-s9y builds started failing with
package-lint-20210326.241, resulting in
```
ox-s9y.el:72:0: error: "org-export-s9y" doesn't start with package's
prefix "ox-s9y".
```

## analysis

`org-export-s9y` refers to the configuration group and the name indeed
different from out package name.  package-lint has any right to throw
an error.

BUT in this case it is ok: In the org-mode source, every exporter that
has configuration options does the same: a filename `ox-ABC.el` has
the configuration group `org-export-ABC` (`ox` is an abbreviation of
`org-export`).

`ox-s9y` wants to keep in line with the org-mode naming conventions so
that its configuration options show up next to the other org-export
groups.

## solution

Use a quick and dirty hack to rename the configuration option before
running package-lint.